Fix bug with count not including all issues
The count attribute returned with the json from Bitbucket does not
include tickets that no longer exist so simply using range(count)
doesn't get all the tickets when creating the issue_ids. Just set the
max_id to 4000. Terrible hack, but I don't want to spend anymore time
on this.